get rid of raw rnn encoder
