+++ tests/audio-classifier.ini
@@ -50,12 +50,14 @@ path="tests/data/dtmf/labels.vocab"
 contains_header=False
 contains_frequencies=False
 
-
-[encoder]
-class=encoders.raw_rnn_encoder.RawRNNEncoder
-name="encoder"
+[input_seq]
+class=encoders.numpy_stateful_filler.TemporalFiller
 data_id="features"
 input_size=26
+
+[encoder]
+class=encoders.RecurrentEncoder
+input_sequence=<input_seq>
 rnn_layers=[(7)]
 dropout_keep_prob=0.5
