+++ tests/ctc.ini
@@ -52,11 +52,14 @@ path="tests/data/yesno/yesno.vocab"
 contains_header=False
 contains_frequencies=False
 
-[audio_encoder]
-class=encoders.raw_rnn_encoder.RawRNNEncoder
-name="audio_encoder"
+[input_seq]
+class=encoders.numpy_stateful_filler.TemporalFiller
 data_id="source"
 input_size=39
+
+[audio_encoder]
+class=encoders.RecurrentEncoder
+input_sequence=<input_seq>
 rnn_layers=[(50,"bidirectional"),(100,"forward"),(100,"backward")]
 dropout_keep_prob=0.5
